Adding c++ equivalent headers & api files
1) no longer re-declaring functions in conv_tuner
2) new source files have same names as their c counterparts, but live
in the tuner folder
3) new sources have all opaque pointers converted to their c++ objects
4) Putting c++ object references in miopen.hpp is tricky. miopen.h was
originally loaded before the declarations of the objects it refers to,
and the declarations of those objects rely on some of the type
structures created in miopen.h. For now this is resolved by moving up
the type declarations in miopen.hpp and including the object
declarations after these structures are defined.